feat(editor-v2): port AI tab into editors/ProductEditorV2 and add ProductEditorV2 wrapper
